renamed context.tsx => ThemeContext.tsx
